Refactor catalog definitions into demos
